remove redundant reset view button on map
